+++ relay-server/src/utils/dynamic_sampling.rs
@@ -5,7 +5,7 @@ use std::net::IpAddr;
 use relay_common::{ProjectKey, Uuid};
 use relay_general::protocol::Event;
 use relay_sampling::{
-    pseudo_random_from_uuid, DynamicSamplingContext, RuleId, SamplingConfig, SamplingRule,
+    pseudo_random_from_uuid, DynamicSamplingContext, RuleId, SamplingConfig, SamplingMode,
 };
 
 use crate::actors::project::ProjectState;
@@ -46,12 +46,19 @@ fn check_unsupported_rules(
     Ok(())
 }
 
-fn get_trace_sampling_rule<'a>(
+#[derive(Clone, Debug, PartialEq)]
+struct SamplingSpec {
+    sample_rate: f64,
+    rule_id: RuleId,
+    seed: Uuid,
+}
+
+fn get_trace_sampling_rule(
     processing_enabled: bool,
-    sampling_project_state: Option<&'a ProjectState>,
+    sampling_project_state: Option<&ProjectState>,
     sampling_context: Option<&DynamicSamplingContext>,
     ip_addr: Option<IpAddr>,
-) -> Result<Option<(&'a SamplingRule, Uuid)>, SamplingResult> {
+) -> Result<Option<SamplingSpec>, SamplingResult> {
     let sampling_context = or_ok_none!(sampling_context);
 
     if sampling_project_state.is_none() {
@@ -62,15 +69,25 @@ fn get_trace_sampling_rule<'a>(
     check_unsupported_rules(processing_enabled, sampling_config)?;
 
     let rule = or_ok_none!(sampling_config.get_matching_trace_rule(sampling_context, ip_addr));
-    Ok(Some((rule, sampling_context.trace_id)))
+    let sample_rate = match sampling_config.mode {
+        SamplingMode::Received => rule.sample_rate,
+        SamplingMode::Total => sampling_context.adjusted_sample_rate(rule.sample_rate),
+    };
+
+    Ok(Some(SamplingSpec {
+        sample_rate,
+        rule_id: rule.id,
+        seed: sampling_context.trace_id,
+    }))
 }
 
-fn get_event_sampling_rule<'a>(
+fn get_event_sampling_rule(
     processing_enabled: bool,
-    project_state: &'a ProjectState,
+    project_state: &ProjectState,
+    sampling_context: Option<&DynamicSamplingContext>,
     event: Option<&Event>,
     ip_addr: Option<IpAddr>,
-) -> Result<Option<(&'a SamplingRule, Uuid)>, SamplingResult> {
+) -> Result<Option<SamplingSpec>, SamplingResult> {
     let event = or_ok_none!(event);
     let event_id = or_ok_none!(event.id.value());
 
@@ -78,7 +95,16 @@ fn get_event_sampling_rule<'a>(
     check_unsupported_rules(processing_enabled, sampling_config)?;
 
     let rule = or_ok_none!(sampling_config.get_matching_event_rule(event, ip_addr));
-    Ok(Some((rule, event_id.0)))
+    let sample_rate = match (sampling_context, sampling_config.mode) {
+        (Some(ctx), SamplingMode::Total) => ctx.adjusted_sample_rate(rule.sample_rate),
+        _ => rule.sample_rate,
+    };
+
+    Ok(Some(SamplingSpec {
+        sample_rate,
+        rule_id: rule.id,
+        seed: event_id.0,
+    }))
 }
 
 /// Checks whether an event should be kept or removed by dynamic sampling.
@@ -98,27 +124,32 @@ pub fn should_keep_event(
         sampling_context,
         ip_addr,
     ) {
-        Ok(rule_and_uuid) => rule_and_uuid,
+        Ok(spec) => spec,
         Err(sampling_result) => return sampling_result,
     };
 
-    let matching_event_rule =
-        match get_event_sampling_rule(processing_enabled, project_state, event, ip_addr) {
-            Ok(rule_and_uuid) => rule_and_uuid,
-            Err(sampling_result) => return sampling_result,
-        };
+    let matching_event_rule = match get_event_sampling_rule(
+        processing_enabled,
+        project_state,
+        sampling_context,
+        event,
+        ip_addr,
+    ) {
+        Ok(spec) => spec,
+        Err(sampling_result) => return sampling_result,
+    };
 
-    for (rule, uuid) in matching_trace_rule.into_iter().chain(matching_event_rule) {
-        let adjusted_sample_rate = if let Some(sampling_context) = sampling_context {
-            sampling_context.adjusted_sample_rate(rule.sample_rate)
+    // NOTE: Event rules take precedence over trace rules. If the event rule has a lower sample rate
+    // than the trace rule, this means that traces will be incomplete.
+    // We could guarantee consistent traces if trace rules took precedence over event rules,
+    // but we need the current behavior to allow health check rules
+    // to take precedence over the overall base rate, which is set on the trace.
+    if let Some(spec) = matching_event_rule.or(matching_trace_rule) {
+        let random_number = pseudo_random_from_uuid(spec.seed);
+        if random_number >= spec.sample_rate {
+            return SamplingResult::Drop(spec.rule_id);
         } else {
-            rule.sample_rate
-        };
-
-        let random_number = pseudo_random_from_uuid(uuid);
-
-        if random_number >= adjusted_sample_rate {
-            return SamplingResult::Drop(rule.id);
+            return SamplingResult::Keep;
         }
     }
 
@@ -139,57 +170,56 @@ pub fn get_sampling_key(envelope: &Envelope) -> Option<ProjectKey> {
 
 #[cfg(test)]
 mod tests {
-    use std::time::Instant;
-
     use bytes::Bytes;
-    use smallvec::SmallVec;
 
     use relay_common::EventType;
     use relay_general::protocol::EventId;
     use relay_general::types::Annotated;
-    use relay_sampling::{RuleId, RuleType, SamplingConfig};
+    use relay_sampling::{RuleCondition, RuleId, RuleType, SamplingConfig, SamplingRule};
 
-    use crate::actors::project::ProjectConfig;
     use crate::envelope::Item;
 
     use super::*;
 
-    fn get_project_state(sample_rate: Option<f64>, rule_type: RuleType) -> ProjectState {
-        let sampling_config_str = if let Some(sample_rate) = sample_rate {
-            let rt = match rule_type {
-                RuleType::Transaction => "transaction",
-                RuleType::Error => "error",
-                RuleType::Trace => "trace",
-            };
-            format!(
-                r#"{{
-                "rules":[{{
-                    "condition": {{ "op": "and", "inner":[]}},
-                    "sampleRate": {},
-                    "type": "{}",
-                    "id": 1
-                }}]
-            }}"#,
-                sample_rate, rt
-            )
-        } else {
-            "{\"rules\":[]}".to_owned()
-        };
-        let sampling_config = serde_json::from_str::<SamplingConfig>(&sampling_config_str).ok();
+    fn state_with_config(sampling_config: SamplingConfig) -> ProjectState {
+        let mut state = ProjectState::allowed();
+        state.config.dynamic_sampling = Some(sampling_config);
+        state
+    }
 
-        ProjectState {
-            project_id: None,
-            disabled: false,
-            public_keys: SmallVec::new(),
-            slug: None,
-            config: ProjectConfig {
-                dynamic_sampling: sampling_config,
-                ..ProjectConfig::default()
-            },
-            organization_id: None,
-            last_change: None,
-            last_fetch: Instant::now(),
-            invalid: false,
+    fn state_with_rule(
+        sample_rate: Option<f64>,
+        rule_type: RuleType,
+        mode: SamplingMode,
+    ) -> ProjectState {
+        let rules = match sample_rate {
+            Some(sample_rate) => vec![SamplingRule {
+                condition: RuleCondition::all(),
+                sample_rate,
+                ty: rule_type,
+                id: RuleId(1),
+                time_range: Default::default(),
+            }],
+            None => Vec::new(),
+        };
+
+        state_with_config(SamplingConfig {
+            rules,
+            mode,
+            next_id: None,
+        })
+    }
+
+    fn create_sampling_context(sample_rate: Option<f64>) -> DynamicSamplingContext {
+        DynamicSamplingContext {
+            trace_id: uuid::Uuid::new_v4(),
+            public_key: "12345678901234567890123456789012".parse().unwrap(),
+            release: None,
+            environment: None,
+            transaction: None,
+            sample_rate,
+            user: Default::default(),
+            other: Default::default(),
         }
     }
 
@@ -199,18 +229,11 @@ mod tests {
         let event_id = EventId::new();
 
         let raw_event = if with_trace_context {
-            let trace_id = uuid::Uuid::new_v4();
-            let project_key = "12345678901234567890123456789012";
-            let trace_context_raw = format!(
-                r#"{{"trace_id": "{}", "public_key": "{}"}}"#,
-                trace_id.to_simple(),
-                project_key,
-            );
             format!(
                 "{{\"event_id\":\"{}\",\"dsn\":\"{}\", \"trace\": {}}}\n",
                 event_id.0.to_simple(),
                 dsn,
-                trace_context_raw,
+                serde_json::to_string(&create_sampling_context(None)).unwrap(),
             )
         } else {
             format!(
@@ -245,18 +268,18 @@ mod tests {
             ..Event::default()
         };
 
-        let proj_state = get_project_state(Some(0.0), RuleType::Error);
+        let proj_state = state_with_rule(Some(0.0), RuleType::Error, SamplingMode::default());
 
         assert_eq!(
             SamplingResult::Drop(RuleId(1)),
             should_keep_event(None, Some(&event), None, &proj_state, None, true)
         );
-        let proj_state = get_project_state(Some(1.0), RuleType::Error);
+        let proj_state = state_with_rule(Some(1.0), RuleType::Error, SamplingMode::default());
         assert_eq!(
             SamplingResult::Keep,
             should_keep_event(None, Some(&event), None, &proj_state, None, true)
         );
-        let proj_state = get_project_state(None, RuleType::Error);
+        let proj_state = state_with_rule(None, RuleType::Error, SamplingMode::default());
         assert_eq!(
             SamplingResult::Keep,
             should_keep_event(None, Some(&event), None, &proj_state, None, true)
@@ -267,8 +290,8 @@ mod tests {
     fn test_unsampled_envelope_with_sample_rate() {
         //create an envelope with a event and a transaction
         let envelope = new_envelope(true);
-        let state = get_project_state(Some(1.0), RuleType::Trace);
-        let sampling_state = get_project_state(Some(0.0), RuleType::Trace);
+        let state = state_with_rule(Some(1.0), RuleType::Trace, SamplingMode::default());
+        let sampling_state = state_with_rule(Some(0.0), RuleType::Trace, SamplingMode::default());
         let result = should_keep_event(
             envelope.sampling_context(),
             None,
@@ -285,8 +308,8 @@ mod tests {
     fn test_should_keep_transaction_no_trace() {
         //create an envelope with a event and a transaction
         let envelope = new_envelope(false);
-        let state = get_project_state(Some(1.0), RuleType::Trace);
-        let sampling_state = get_project_state(Some(0.0), RuleType::Trace);
+        let state = state_with_rule(Some(1.0), RuleType::Trace, SamplingMode::default());
+        let sampling_state = state_with_rule(Some(0.0), RuleType::Trace, SamplingMode::default());
 
         let result = should_keep_event(
             envelope.sampling_context(),
@@ -307,8 +330,8 @@ mod tests {
     fn test_should_signal_when_envelope_becomes_empty() {
         //create an envelope with a event and a transaction
         let envelope = new_envelope(true);
-        let state = get_project_state(Some(1.0), RuleType::Trace);
-        let sampling_state = get_project_state(Some(0.0), RuleType::Trace);
+        let state = state_with_rule(Some(1.0), RuleType::Trace, SamplingMode::default());
+        let sampling_state = state_with_rule(Some(0.0), RuleType::Trace, SamplingMode::default());
 
         let result = should_keep_event(
             envelope.sampling_context(),
@@ -319,5 +342,142 @@ mod tests {
             true,
         );
         assert_eq!(result, SamplingResult::Drop(RuleId(1)));
+    }
+
+    #[test]
+    /// When there's a mixture of event rules and trace rules, the event rules
+    /// take precedence.
+    fn test_event_rule_precedence() {
+        let sampling_config = serde_json::json!(
+            {
+                "rules": [
+                    {
+                        "sampleRate": 0,
+                        "type": "trace",
+                        "active": true,
+                        "condition": {
+                            "op": "and",
+                            "inner": []
+                        },
+                        "id": 1000
+                    },
+                    {
+                        "sampleRate": 1,
+                        "type": "transaction",
+                        "condition": {
+                            "op": "or",
+                            "inner": [
+                            {
+                                "op": "glob",
+                                "name": "event.transaction",
+                                "value": [
+                                    "my-important-transaction",
+                                ],
+                                "options": {
+                                    "ignoreCase": true
+                                }
+                            }
+                            ]
+                        },
+                        "active": true,
+                        "id": 1002
+                    }
+                ]
+            }
+        );
+
+        let sampling_config = serde_json::from_value(sampling_config).unwrap();
+        let project_state = state_with_config(sampling_config);
+
+        let envelope = new_envelope(true);
+
+        let event = Event {
+            id: Annotated::new(EventId::new()),
+            ty: Annotated::new(EventType::Transaction),
+            transaction: Annotated::new("my-important-transaction".to_owned()),
+            ..Event::default()
+        };
+
+        let keep_event = should_keep_event(
+            envelope.sampling_context(),
+            Some(&event),
+            None,
+            &project_state,
+            Some(&project_state),
+            true,
+        );
+
+        assert_eq!(keep_event, SamplingResult::Keep);
+    }
+
+    #[test]
+    fn test_trace_rule_received() {
+        let project_state = state_with_rule(Some(0.1), RuleType::Trace, SamplingMode::Received);
+        let sampling_context = create_sampling_context(Some(0.5));
+        let spec = get_trace_sampling_rule(
+            true, // irrelevant, just skips unsupported rules
+            Some(&project_state),
+            Some(&sampling_context),
+            None,
+        );
+
+        assert_eq!(spec.unwrap().unwrap().sample_rate, 0.1);
+    }
+
+    #[test]
+    fn test_trace_rule_adjusted() {
+        let project_state = state_with_rule(Some(0.1), RuleType::Trace, SamplingMode::Total);
+        let sampling_context = create_sampling_context(Some(0.5));
+        let spec = get_trace_sampling_rule(
+            true, // irrelevant, just skips unsupported rules
+            Some(&project_state),
+            Some(&sampling_context),
+            None,
+        );
+
+        assert_eq!(spec.unwrap().unwrap().sample_rate, 0.2);
+    }
+
+    #[test]
+    fn test_event_rule_received() {
+        let project_state =
+            state_with_rule(Some(0.1), RuleType::Transaction, SamplingMode::Received);
+        let sampling_context = create_sampling_context(Some(0.5));
+        let event = Event {
+            id: Annotated::new(EventId::new()),
+            ty: Annotated::new(EventType::Transaction),
+            ..Event::default()
+        };
+
+        let spec = get_event_sampling_rule(
+            true, // irrelevant, just skips unsupported rules
+            &project_state,
+            Some(&sampling_context),
+            Some(&event),
+            None, // ip address not needed for uniform rule
+        );
+
+        assert_eq!(spec.unwrap().unwrap().sample_rate, 0.1);
+    }
+
+    #[test]
+    fn test_event_rule_adjusted() {
+        let project_state = state_with_rule(Some(0.1), RuleType::Transaction, SamplingMode::Total);
+        let sampling_context = create_sampling_context(Some(0.5));
+        let event = Event {
+            id: Annotated::new(EventId::new()),
+            ty: Annotated::new(EventType::Transaction),
+            ..Event::default()
+        };
+
+        let spec = get_event_sampling_rule(
+            true, // irrelevant, just skips unsupported rules
+            &project_state,
+            Some(&sampling_context),
+            Some(&event),
+            None, // ip address not needed for uniform rule
+        );
+
+        assert_eq!(spec.unwrap().unwrap().sample_rate, 0.2);
     }
 }